Remove cyclonedx task dependency on do_fetch
All relevant information for constructing the sbom is available within
the recipes themselves, thus there is no reason to depend on do_fetch.

Removing this dependency has the advantage that when running bitbake
with the --runonly=do_cyclonedx_package_collect flag (i.e. one is interested
in only the sbom, not actual build artifacts) will not depend on
unnecessary fetch tasks, thus generating the sbom will be fast
regardless of the availability of a populated dl_dir.